Split tasks into files
