+++ src/identity_committer/tasks/mod.rs
@@ -1,0 +1,2 @@
+mod mine_identities;
+mod process_identities;
